+++ lib/instagram/getPostComments.ts
@@ -2,7 +2,7 @@ import { Post } from "../../types/agent";
 import getActorStatus from "../apify/getActorStatus";
 import getDataset from "../apify/getDataset";
 import getFormattedComments from "./getFormattedComments";
-import startCommentsScraping from "./getPostCommentsDatasetId";
+import startCommentsScraping from "./startCommentsScraping";
 
 const MAX_ATTEMPTS = 30;
 const POLLING_INTERVAL = 3000; // 3 seconds
